pyproject.toml
[build-system]
requires = [ "setuptools>=77.0", "setuptools_scm[toml]>=8.0",]
build-backend = "setuptools.build_meta"

[project]
name = "bmm_instrument"
version = "0.0.1"
description = "BMM NSLS-II BITS Instrument Package"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["NSLS-II", "BMM", "XAFS", "beamline", "bluesky"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "Programming Language :: Python :: 3.11",
]
license = "BSD-3-Clause"
license-files = ["LICENSE*"]
dependencies = [
    "apsbits",
    "ophyd-async>=0.3.0",
    "bluesky>=1.10.0",
    "databroker",
    "numpy",
    "matplotlib",
    "pandas",
    "pyyaml",
]

authors = [
    { name = "Rafael Vescovi", email = "rvescovi@anl.gov" },
]
maintainers = [
    { name = "Rafael Vescovi", email = "rvescovi@anl.gov" },
]

[tool.copyright]
copyright = "2025, NSLS-II BMM Beamline"

[tool.black]
line-length = 115
target-version = [ "py311",]
include = "\\.pyi?$"

[tool.flake8]
max-line-length = 115
extend-ignore = [ "E203", "E402", "E501", "F401", "F405", "F722", "F811", "W503", "W504",]

[tool.isort]
force_single_line = true
line_length = 88
include_trailing_comma = true

[tool.ruff]
exclude = [ ".eggs", ".git", ".git-rewrite", ".mypy_cache", ".pytype", ".ruff_cache", ".tox", ".venv", "__pypackages__", "_build", "build", "dist", "venv", ".venv", "docs",]
line-length = 88
indent-width = 4
target-version = "py311"

[project.optional-dependencies]
dev = [ "build", "isort", "mypy", "pre-commit", "pytest", "ruff",]
doc = [ "babel", "ipykernel", "jinja2", "markupsafe", "myst_parser", "nbsphinx", "pydata-sphinx-theme", "pygments-ipython-console", "pygments", "sphinx-design", "sphinx-tabs", "sphinx",]
all = [ "bmm_instrument[dev,doc]",]

[project.urls]
Homepage = "https://github.com/ravescovi/bmm-nsls-bits"
"Bug Tracker" = "https://github.com/ravescovi/bmm-nsls-bits/issues"

[tool.pytest.ini_options]
addopts = [ "--import-mode=importlib", "-x",]
junit_family = "xunit1"
filterwarnings = [ "ignore::DeprecationWarning", "ignore::PendingDeprecationWarning",]

[tool.ruff.lint]
select = [ "E", "F", "B", "I", "W", "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",]
fixable = [ "ALL",]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
"*" = [ "*.yml",]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = [ "F405",]

[tool.ruff.lint.isort]
force-single-line = true
